wechat-chat-manager/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# WeChat Memory Reading
pymem>=1.13.0

# Cryptography
pycryptodome>=3.19.0

# Note: SQLCipher decryption handled by pycryptodome
# sqlcipher3-binary not available on Windows

# Secure credential storage
keyring>=24.0.0

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Windows process utilities
psutil>=5.9.0
